eturnal_ctl: Don't log errors
The control command API is meant to be used interactively, so errors
returned by eturnal_ctl functions are usually displayed to the user.
Log messages would be printed to his console as well (except for 'info'
and 'debug' messages, which OTP hides by default).  Therefore, error
messages would be duplicated.