Don't abort if TURN is enabled without 'secret'
The idea behind this change is making the example configuration more
suitable for the (presumably common) case where TURN support is desired.

The shipped configuration file no longer disables TURN.  As long as no
authentication secret is specified (or the relay address is unknown),
eturnal will now implicitly disable TURN and log an appropriate warning,
but it won't abort.  To enable TURN, a 'secret' (and possibly the
'relay_ipv4_addr') must be specified as mentioned in the warnings, but
it's no longer necessary to remove the 'enable_turn: false' lines from
the example configuration.

If TURN support is actually undesired, 'enable_turn: false' lines can be
specified in order to avoid the warnings.