Support multiplexing TCP/TLS listeners
If the 'listen' option 'transport' is set to 'auto', accept both
unencrypted TCP and TLS connections on the same socket.

This feature depends on the new 'socket' backend for TCP/TLS
connections, and therefore on Erlang/OTP 23 or later.

Many thanks to Annika Hannig for her work on this functionality.