eturnalctl: Read eturnalctl.cfg file
Allow for specifying a few startup configuration options in an
eturnalctl.cfg file.  Leave this functionality undocumented for the
moment, as it's probably preferable to stick to a single configuration
file for eturnal.  However, now that the binary release sets the
ERL_DIST_PORT variable by default, it might be good to be able to offer
some solution for adjusting it (which doesn't involve editing the
eturnalctl script directly).