Optionally accept expired credentials for refresh
If the new 'strict_expiry' option is set to 'true', expired credentials
won't be accepted for refreshing existing TURN sessions.  This is how
eturnal behaved until now.

If the option is set to 'false' (which is the default, now), the expiry
timestamp of ephemeral credentials is only checked during TURN session
creation.  Once credentials are accepted, clients may continue to use
them for refreshing the existing session beyond credential expiry.  This
makes sure the credential expiry timestamp doesn't limit the call time.
Coturn currently behaves this way as well, see:

https://github.com/coturn/coturn/blob/4.5.2/README.turnserver#L750-L757
https://github.com/coturn/coturn/issues/774