Dockerfile: Adjust package installation commands
Change package upgrade to only upgrade available packages and not to reset to available/latest packages in repositories. Herewith, packages are not removed which are included in the base image.

APK cache is deleted anyway with the removal of the apk-tool package, so no need to specific it in the apk commands anymore.