Dockerfile: Improve compatibility between Alpine versions
This may be beneficial if one specifies an OTP_VSN which was build on an older Alpine version base, e.g. 3.16.

Background: the busybox package was split in vsn 3.17 into busybox and busybox-binsh. busybox-binsh is basically only a symlink to busybox which we just adopted. Now we just mention busybox as a general runtime requirement without the hassle, that in versions 3.17 onwards the busybox-binsh package must be defined to avoid deleting the ash symlink to busybox. This would, however, break the docker build in versions < 3.17 as busybox-binsh does not exist in the Alpine repository.

Alternatively one could also use something like this:
RUN check=$(printf "$(cat /etc/alpine-release)\n3.17" | sort -V | head -n1) \
    && if [ "$check" = '3.17' ]; then sub='-binsh'; fi \
    && apk add --no-cache \
        busybox${sub}