Dockerfile: use a specific erlang/otp vsn to compile
We use 'docker.io/erlang' as basis and cleanup the image to use for runtime. This avoids any compatibility issue which may arise with differing Alpine versions if using a simple Alpine image as basis for runtime.

For Dockerfile maintanence it is preferred to use a specific erlang/otp vsn. A release, e.g.'25.3' should be the hypothetical tag '25.3.0.0'. This keeps the alpine base image the same and avoids errors with changes in package names. The 'docker.io/erlang' repo is: https://github.com/erlang/docker-erlang-otp

There is no image avialable for arm/v6, so we will drop this architecture.