eturnalctl: Use readlink(1)'s "-f" option
Let readlink(1) canonicalize the entire path by following every
component, and let it always print a result on BSD-style systems as
well.

The readlink(1) utility is not specified by POSIX, and the "-f" option
isn't available on all systems.  However, it's supported by GNU
readlink, on FreeBSD, NetBSD, and OpenBSD.  The option has also been
added to macOS 12.3.  If we ever need support for a system that doesn't
have it, we should do something like:

https://github.com/ko1nksm/readlinkf