publish-release: Simplify check for musl
glibc's libc.so path name doesn't necessarily contain the "gnu" string.
We could execute the libc.so file and check the output (this is
supported by both glibc and musl). But let's keep it simple and see how
far we get by blindly assuming glibc if the ldd output doesn't contain
the "musl" string.