Let status command check whether eturnal is ready
Add a custom command extension to let "eturnalctl status" check whether
eturnal is actually ready to handle STURN/TURN clients (and print a line
to the standard output in that case).  If this call is issued early
during startup, it will block (up to 15 seconds) until eturnal is
responsive.  The old behavior was to rely on Relx' default
implementation of the "status" command, which (silently) returns success
as long as the Erlang node is alive.